requirements.txt
openai>=1.30.0
gspread>=6.0.0
google-auth>=2.29.0
boto3>=1.34.0
flask>=3.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
